lambda/document-generator/requirements.txt
anthropic==0.39.0
python-docx==1.1.0
